requirements.txt
uvicorn~=0.24.0
fastapi~=0.116.1
slack_sdk~=3.36.0
dotenv~=0.9.9
python-dotenv~=1.1.0
openai~=1.88.0
pdf2image~=1.17.0
Pillow~=10.1.0
PyMuPDF~=1.23.8
reportlab~=4.4.2